add reverse symlinked files.
