--- fairseq_cli/setup.py
@@ -1,1 +1,0 @@
-../setup.py
