+++ train.py
@@ -1,296 +1,1 @@
-#!/usr/bin/env python3 -u
-# Copyright (c) Facebook, Inc. and its affiliates.
-#
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-"""
-Train a new model on one or across multiple GPUs.
-"""
-
-import math
-import random
-
-import numpy as np
-import torch
-
-from fairseq import (
-    checkpoint_utils, distributed_utils, metrics, options, progress_bar, tasks, utils
-)
-from fairseq.data import iterators
-from fairseq.trainer import Trainer
-from fairseq.meters import StopwatchMeter
-
-
-def main(args, init_distributed=False):
-    utils.import_user_module(args)
-
-    assert args.max_tokens is not None or args.max_sentences is not None, \
-        'Must specify batch size either with --max-tokens or --max-sentences'
-
-    # Initialize CUDA and distributed training
-    if torch.cuda.is_available() and not args.cpu:
-        torch.cuda.set_device(args.device_id)
-    np.random.seed(args.seed)
-    torch.manual_seed(args.seed)
-    if init_distributed:
-        args.distributed_rank = distributed_utils.distributed_init(args)
-
-    if distributed_utils.is_master(args):
-        checkpoint_utils.verify_checkpoint_directory(args.save_dir)
-
-    # Print args
-    print(args)
-
-    # Setup task, e.g., translation, language modeling, etc.
-    task = tasks.setup_task(args)
-
-    # Load valid dataset (we load training data below, based on the latest checkpoint)
-    for valid_sub_split in args.valid_subset.split(','):
-        task.load_dataset(valid_sub_split, combine=False, epoch=0)
-
-    # Build model and criterion
-    model = task.build_model(args)
-    criterion = task.build_criterion(args)
-    print(model)
-    print('| model {}, criterion {}'.format(args.arch, criterion.__class__.__name__))
-    print('| num. model params: {} (num. trained: {})'.format(
-        sum(p.numel() for p in model.parameters()),
-        sum(p.numel() for p in model.parameters() if p.requires_grad),
-    ))
-
-    # Build trainer
-    trainer = Trainer(args, task, model, criterion)
-    print('| training on {} GPUs'.format(args.distributed_world_size))
-    print('| max tokens per GPU = {} and max sentences per GPU = {}'.format(
-        args.max_tokens,
-        args.max_sentences,
-    ))
-
-    # Load the latest checkpoint if one is available and restore the
-    # corresponding train iterator
-    extra_state, epoch_itr = checkpoint_utils.load_checkpoint(args, trainer)
-
-    # Train until the learning rate gets too small
-    max_epoch = args.max_epoch or math.inf
-    max_update = args.max_update or math.inf
-    lr = trainer.get_lr()
-    train_meter = StopwatchMeter()
-    train_meter.start()
-    valid_subsets = args.valid_subset.split(',')
-    while (
-        lr > args.min_lr
-        and (
-            epoch_itr.epoch < max_epoch
-            # allow resuming training from the final checkpoint
-            or epoch_itr._next_epoch_itr is not None
-        )
-        and trainer.get_num_updates() < max_update
-    ):
-        # train for one epoch
-        train(args, trainer, task, epoch_itr)
-
-        if not args.disable_validation and epoch_itr.epoch % args.validate_interval == 0:
-            valid_losses = validate(args, trainer, task, epoch_itr, valid_subsets)
-        else:
-            valid_losses = [None]
-
-        # only use first validation loss to update the learning rate
-        lr = trainer.lr_step(epoch_itr.epoch, valid_losses[0])
-
-        # save checkpoint
-        if epoch_itr.epoch % args.save_interval == 0:
-            checkpoint_utils.save_checkpoint(args, trainer, epoch_itr, valid_losses[0])
-
-        # early stop
-        if should_stop_early(args, valid_losses[0]):
-            print('| Early stop since valid performance hasn\'t improved for last {} runs'.format(args.patience))
-            break
-
-        reload_dataset = ':' in getattr(args, 'data', '')
-        # sharded data: get train iterator for next epoch
-        epoch_itr = trainer.get_train_iterator(epoch_itr.epoch, load_dataset=reload_dataset)
-    train_meter.stop()
-    print('| done training in {:.1f} seconds'.format(train_meter.sum))
-
-
-def should_stop_early(args, valid_loss):
-    if args.patience <= 0:
-        return False
-
-    def is_better(a, b):
-        return a > b if args.maximize_best_checkpoint_metric else a < b
-
-    prev_best = getattr(should_stop_early, 'best', None)
-    if prev_best is None or is_better(valid_loss, prev_best):
-        should_stop_early.best = valid_loss
-        should_stop_early.num_runs = 0
-        return False
-    else:
-        should_stop_early.num_runs += 1
-        return should_stop_early.num_runs > args.patience
-
-
-def train(args, trainer, task, epoch_itr):
-    """Train the model for one epoch."""
-    # Initialize data iterator
-    itr = epoch_itr.next_epoch_itr(
-        fix_batches_to_gpus=args.fix_batches_to_gpus,
-        shuffle=(epoch_itr.epoch >= args.curriculum),
-    )
-    update_freq = (
-        args.update_freq[epoch_itr.epoch - 1]
-        if epoch_itr.epoch <= len(args.update_freq)
-        else args.update_freq[-1]
-    )
-    itr = iterators.GroupedIterator(itr, update_freq)
-    progress = progress_bar.build_progress_bar(
-        args, itr, epoch_itr.epoch, no_progress_bar='simple',
-    )
-
-    valid_subsets = args.valid_subset.split(',')
-    max_update = args.max_update or math.inf
-    for samples in progress:
-        with metrics.aggregate('train_inner'):
-            log_output = trainer.train_step(samples)
-            num_updates = trainer.get_num_updates()
-            if log_output is None:
-                continue
-
-            # log mid-epoch stats
-            stats = get_training_stats('train_inner')
-            progress.log(stats, tag='train', step=num_updates)
-
-        if (
-            not args.disable_validation
-            and args.save_interval_updates > 0
-            and num_updates % args.save_interval_updates == 0
-            and num_updates > 0
-        ):
-            valid_losses = validate(args, trainer, task, epoch_itr, valid_subsets)
-            checkpoint_utils.save_checkpoint(args, trainer, epoch_itr, valid_losses[0])
-
-        if num_updates >= max_update:
-            break
-
-    # log end-of-epoch stats
-    stats = get_training_stats('train')
-    progress.print(stats, tag='train', step=num_updates)
-
-    # reset epoch-level meters
-    metrics.reset_meters('train')
-
-
-def get_training_stats(stats_key):
-    stats = metrics.get_smoothed_values(stats_key)
-    if 'nll_loss' in stats and 'ppl' not in stats:
-        stats['ppl'] = utils.get_perplexity(stats['nll_loss'])
-    stats['wall'] = round(metrics.get_meter('default', 'wall').elapsed_time, 0)
-    return stats
-
-
-def validate(args, trainer, task, epoch_itr, subsets):
-    """Evaluate the model on the validation set(s) and return the losses."""
-
-    if args.fixed_validation_seed is not None:
-        # set fixed seed for every validation
-        utils.set_torch_seed(args.fixed_validation_seed)
-
-    valid_losses = []
-    for subset in subsets:
-        # Initialize data iterator
-        itr = task.get_batch_iterator(
-            dataset=task.dataset(subset),
-            max_tokens=args.max_tokens_valid,
-            max_sentences=args.max_sentences_valid,
-            max_positions=utils.resolve_max_positions(
-                task.max_positions(),
-                trainer.get_model().max_positions(),
-            ),
-            ignore_invalid_inputs=args.skip_invalid_size_inputs_valid_test,
-            required_batch_size_multiple=args.required_batch_size_multiple,
-            seed=args.seed,
-            num_shards=args.distributed_world_size,
-            shard_id=args.distributed_rank,
-            num_workers=args.num_workers,
-        ).next_epoch_itr(shuffle=False)
-        progress = progress_bar.build_progress_bar(
-            args, itr, epoch_itr.epoch,
-            prefix='valid on \'{}\' subset'.format(subset),
-            no_progress_bar='simple'
-        )
-
-        # reset validation loss meters
-        metrics.reset_meters('valid')
-
-        for sample in progress:
-            trainer.valid_step(sample)
-
-        # log validation stats
-        stats = get_valid_stats(args, trainer)
-        progress.print(stats, tag=subset, step=trainer.get_num_updates())
-
-        valid_losses.append(stats[args.best_checkpoint_metric])
-    return valid_losses
-
-
-def get_valid_stats(args, trainer):
-    stats = metrics.get_smoothed_values('valid')
-    if 'nll_loss' in stats and 'ppl' not in stats:
-        stats['ppl'] = utils.get_perplexity(stats['nll_loss'])
-    stats['num_updates'] = trainer.get_num_updates()
-    if hasattr(checkpoint_utils.save_checkpoint, 'best'):
-        key = 'best_{0}'.format(args.best_checkpoint_metric)
-        best_function = max if args.maximize_best_checkpoint_metric else min
-        stats[key] = best_function(
-            checkpoint_utils.save_checkpoint.best,
-            stats[args.best_checkpoint_metric],
-        )
-    return stats
-
-
-def distributed_main(i, args, start_rank=0):
-    args.device_id = i
-    if args.distributed_rank is None:  # torch.multiprocessing.spawn
-        args.distributed_rank = start_rank + i
-    main(args, init_distributed=True)
-
-
-def cli_main():
-    parser = options.get_training_parser()
-    args = options.parse_args_and_arch(parser)
-
-    if args.distributed_init_method is None:
-        distributed_utils.infer_init_method(args)
-
-    if args.distributed_init_method is not None:
-        # distributed training
-        if torch.cuda.device_count() > 1 and not args.distributed_no_spawn:
-            start_rank = args.distributed_rank
-            args.distributed_rank = None  # assign automatically
-            torch.multiprocessing.spawn(
-                fn=distributed_main,
-                args=(args, start_rank),
-                nprocs=torch.cuda.device_count(),
-            )
-        else:
-            distributed_main(args.device_id, args)
-    elif args.distributed_world_size > 1:
-        # fallback for single node with multiple GPUs
-        assert args.distributed_world_size <= torch.cuda.device_count()
-        port = random.randint(10000, 20000)
-        args.distributed_init_method = 'tcp://localhost:{port}'.format(port=port)
-        args.distributed_rank = None  # set based on device id
-        if max(args.update_freq) > 1 and args.ddp_backend != 'no_c10d':
-            print('| NOTE: you may get faster training with: --ddp-backend=no_c10d')
-        torch.multiprocessing.spawn(
-            fn=distributed_main,
-            args=(args, ),
-            nprocs=args.distributed_world_size,
-        )
-    else:
-        # single GPU training
-        main(args)
-
-
-if __name__ == '__main__':
-    cli_main()
+fairseq_cli/train.py
